+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Added
   * AZ-571: Add pre-commit config hook
   * AZ-571: Check for existing Azure SP before trying to create a new one
+  * AZ-571: Ask the user how long the SP password/secret is valid (in years)
 
 Fixed
   * AZ-571: Change command to retrieve Tenant name/domain